+++ src/app/Playground/MainPage.xaml.cs
@@ -3,6 +3,7 @@ using DIPS.Mobile.UI.Resources.Icons;
 using Playground.EirikSamples;
 using Playground.HåvardSamples;
 using Playground.SanderSamples;
+using Playground.SystemMessageSamples;
 using Playground.VetleSamples;
 using Playground.VetleSamples.CollectionViewTests;
 
@@ -120,5 +121,10 @@ public partial class MainPage
         var scannerPage = new SamplingPatientScanRepro();
         var navigationPage = new NavigationPage(scannerPage);
         Shell.Current.Navigation.PushModalAsync(navigationPage);
+    }
+
+    private void GoToSystemMessageReadonlySample(object sender, EventArgs e)
+    {
+        Shell.Current.Navigation.PushAsync(new SystemMessageReadonlySample());
     }
 }